refactor(BREAKING-CHAKGE): Addede trpc, reorganized file structure
