+++ package.json
@@ -21,6 +21,10 @@
     "@hookform/resolvers": "^2.8.10",
     "@next-auth/prisma-adapter": "^1.0.3",
     "@prisma/client": "^3.14.0",
+    "@trpc/client": "^9.25.2",
+    "@trpc/next": "^9.25.2",
+    "@trpc/react": "^9.25.2",
+    "@trpc/server": "^9.25.2",
     "bcrypt": "^5.0.1",
     "cookie": "^0.5.0",
     "framer-motion": "^6",
@@ -32,6 +36,8 @@
     "react-dom": "^18.1.0",
     "react-hook-form": "^7.31.3",
     "react-icons": "^4.3.1",
+    "react-query": "^3.39.1",
+    "superjson": "^1.9.1",
     "swr": "^1.3.0",
     "yup": "^0.32.11",
     "zustand": "^4.0.0-rc.1"
