--- src/components-common/index.ts
@@ -1,11 +1,0 @@
-/**
- * Elements in components-common are meant to be agnostic to the UI framework.
- * Instead of importing the component directly from the package, we will add
- * a wrapper so it can be changed out if need be.
- */
-
-
-export * from './Typography'
-export * from './Card'
-
-export * from './login-btn'
