--- src/components-core/auth/index.tsx
@@ -1,3 +1,0 @@
-export { SignInContainer } from './signin-container';
-export { SignUpContainer } from './signup-container';
-export { AuthForm } from './auth-form';
