--- src/components-core/profile/index.ts
@@ -1,1 +1,0 @@
-export * from './profile-container'
