--- src/components-feat/index.ts
@@ -1,4 +1,0 @@
-export * from './Determine'
-export * from './ErrorBoundary'
-export * from './Layout'
-export * from './ResourceNotFound'
