+++ src/components/Layout.tsx
@@ -1,4 +1,4 @@
-import { LoginButton } from '@components-common';
+import { LoginButton } from '@components';
 import { useSession } from 'next-auth/react';
 import type { ReactNode } from 'react';
 import Image from 'next/image';
