+++ src/components/comps-auth/index.tsx
@@ -1,0 +1,1 @@
+export { AuthForm } from './auth-form';
