+++ src/components/comps-posts/index.ts
@@ -1,6 +1,3 @@
-export * from './post-detail-container'
-export * from './post-create-container'
-export * from './posts-container'
 export * from './post-header'
 export * from './contribution-card'
 export * from './contribution-list'
