+++ src/lib/prisma.ts
@@ -1,3 +1,20 @@
-import { PrismaClient } from "@prisma/client";
 
-export default new PrismaClient();
+
+import { PrismaClient } from '@prisma/client';
+
+const prismaGlobal = global as typeof global & {
+  prisma?: PrismaClient;
+};
+
+const prisma: PrismaClient =
+  prismaGlobal.prisma ||
+  new PrismaClient({
+    log:
+      process.env.NODE_ENV === 'development' ? ['query', 'error', 'warn'] : ['error'],
+  });
+
+if (process.env.NODE_ENV !== 'production') {
+  prismaGlobal.prisma = prisma;
+}
+
+export default prisma
