+++ src/lib/trpc.ts
@@ -1,0 +1,6 @@
+import { createReactQueryHooks } from '@trpc/react';
+import { AppRouter } from 'server/routers/_app';
+
+
+export const trpc = createReactQueryHooks<AppRouter>();
+// => { useQuery: ..., useMutation: ...}
