+++ src/lib/usePost.ts
@@ -1,44 +1,27 @@
-import { Block, Contribution, Post, User } from '@prisma/client';
-import fetcher from './fetcher';
-import useSWR from 'swr';
+import { useToast } from "@chakra-ui/react";
+import { Post } from "@prisma/client";
+import { useRouter } from "next/router";
+import { trpc } from "./trpc";
 
-interface UserType extends Contribution {
-  author: Pick<User, 'name'>;
-}
-
-interface BlockType extends Block {
-  contributions: UserType[];
-  _count: {
-    contributions: Number;
-  };
-}
-export interface UsePostTypes extends Post {
-  author: User;
-  blocks: BlockType[];
-}
-
-
-// The api will send back {error: true, message: "Post not found"} if user navigates to a post detail page which does not exist on a dynamic route
-const hasData = <T>(data: T ): T | undefined => {
-  if(!data){
-    return undefined
-  }
-  if(typeof data === "object" && 'error' in data){
-    return undefined;
-  }
-  return data
-};
-
-export const usePost = (post_uuid: Post["post_uuid"]) => {
-  const { data, error, mutate } = useSWR<UsePostTypes>(
-    post_uuid ? `/posts/${post_uuid}` : null,
-    fetcher,
-  );
-
-  return {
-    post: hasData(data),
-    loading: !data && !error,
-    error: error,
-    mutate,
+export const usePost = () => {
+  const toast = useToast()
+  const router = useRouter()
+ return {
+    getPost: (post_uuid: Post["post_uuid"]) => trpc.useQuery(['posts.byId', {post_uuid}]),
+    deletePost: () => {},
+    createPost: (options: {onSuccesFunc: () => void }) =>  trpc.useMutation(['posts.create'], {
+      onSuccess: data => {
+        options.onSuccesFunc()
+        router.push(`/posts/${data?.post_uuid}`);
+      },
+      onError: error => {
+        toast({
+          position: 'top',
+          status: 'error',
+          description: error.message,
+          isClosable: true,
+        });
+      },
+    })
   };
 };
