+++ src/pages/api/auth/[...nextauth].ts
@@ -1,4 +1,4 @@
-import NextAuth from "next-auth"
+import NextAuth, { Session } from "next-auth"
 import GoogleProvider from "next-auth/providers/google"
 import { PrismaAdapter } from "@next-auth/prisma-adapter"
 import { PrismaClient } from "@prisma/client"
@@ -34,3 +34,6 @@ export default NextAuth({
   }
 })
 
+export interface NewSession extends Session {
+  id?: number
+}
