--- src/pages/posts/index.ts
@@ -1,1 +1,0 @@
-export {PostsContainer as default} from '@components-core/posts'
